week_10/day_1/wk8dailychallenge/requirements.txt
django
django-phonenumber-field[phonenumberslite]